Professor Dashboard (Dashboard and Professors Tab)
Used a tab selector and drawer from material skin 2 package. The tab selector's background has an unchangeable ugly gray color that does not fit with the aesthetic of the app, so I just slapped a card in to make it appear that it changed to white.

Also, kind of lazy to use the material skin's list view control instead of DataGridView because it will change the WHOLE program in setting the table from the data in the database.