Student Login Some Features
Added a student login.

A student validation if they exist on the database and if they're not.

The student name label gets set when logging in.

The date and time label is set too.

The big section label is set too, with the section of the logged in student.

Sign out button is also now functioning.

Also added a focus method when choosing student as user.

Re-designed Student Dashboard, put the navigation bar on top instead of the left side.

Also put a warning message after the submit button, to prompt the user if they are sure with their changes.